+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0
 ### Feature
 * add interactive stop of background jobs
+* introduce git-style commands system
 ### Bug fix
 * fix handling kill of multiple commands run in background